Fix a crash when an issue is recorded on a detached task. (#592)
Given the following code:

```swift
@Test func f() async {
  await Task.detached {
    #expect(1 == 2)
  }.value
}
```

An issue is recorded on a detached task, but Swift Testing is unable to
determine at runtime which test it corresponds to. This should simply
result in the issue being attributed to an arbitrary test, or to the
aether, but #421 caused a regression that triggers an optional unwrap
crash instead.

This PR fixes the crash. :)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.