Implement an overload of `confirmation()` that takes an unbounded range.
Unbounded ranges are not meaningful when used with `confirmation()` because
_any_ confirmation count matches. As well, the `...` operator produces an
instance of `UnboundedRange` which is a non-nominal type and cannot conform to
`RangeExpression`. It may be non-obvious to a developer why `...` doesn't work
as the `expectedCount` argument to that function when other range operators work
as expected.

This PR implements a stub overload of `confirmation()` that takes an unbounded
range. The stub overload is marked unavailable and cannot be called.