Implement an overload of `confirmation()` that takes an unbounded range. (#598)
Unbounded ranges are not meaningful when used with `confirmation()`
because _any_ confirmation count matches. As well, the `...` operator
produces an instance of `UnboundedRange` which is a non-nominal type and
cannot conform to `RangeExpression`. It may be non-obvious to a
developer why `...` doesn't work as the `expectedCount` argument to that
function when other range operators work as expected.

This PR implements a stub overload of `confirmation()` that takes an
unbounded range. The stub overload is marked unavailable and cannot be
called.

Example usage:

```swift
await confirmation("Stuff happens", expectedCount: ...) { stuff in
  // ...
}
```

Generated diagnostic:

> 🛑 'confirmation(\_:expectedCount:sourceLocation:\_:)' is unavailable:
Unbounded range '...' has no effect when used with a confirmation.

As a reminder, using a range expression with `confirmation()` is an
experimental feature and has not been API-reviewed.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.