Fix a merge conflict on main-next.
This PR fixes a conflict between #610 and #619. #610 added a string cache to
`TypeInfo` using `ObjectIdentifier` instances as keys. #619 added support for
move-only types to `TypeInfo`. Due to rdar://134276458, move-only types cannot
be used with `ObjectIdentifier`. This PR uses `UInt` instead until that issue
can be resolved in a future stdlib update.