Disallow `@Test` on member functions of `XCTestCase` subclasses. (#626)
This PR adds a diagnostic if we detect that `@Test` has been used on a
member function of an `XCTestCase` subclass. This was meant to be
disallowed, but we neglected to add a check after adopting
`lexicalContext`.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.