Set the library version reported via CMake to 6.1.0-dev.
This PR updates the version reported when using CMake (i.e. when building in the
toolchain) to 6.1.0-dev with optional trailing git status information, if
available. The resulting string looks like:

> 6.1.0-dev (0.0.0-initial-532-gd01c406 - modified)

"0.0.0-initial-532-gd01c406" in this example refers to the most recent commit
on the branch I was testing on via `git describe --tags`, and "modified" is
present because I had uncommitted changes.

(We'll want to double-check that CMake used in CI can see the git repository and
has access to the git tool.)