Set the library version reported via CMake to 6.1-dev. (#637)
This PR updates the version reported when using CMake (i.e. when
building in the toolchain) to `"6.1-dev"` with optional trailing git
commit hash, if available. The resulting string looks like:

> `6.1.0-dev (c6450e02cc76cd7 - modified)`

"c6450e02cc76cd7" in this example refers to the most recent commit on
the branch I was testing on via `git rev-parse --verify HEAD`, and
"modified" is present because I had uncommitted changes.

(We'll want to double-check that CMake used in CI can see the git
repository and has access to the git tool.)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.