Plumb through the testing library version into a function instead of a macro.
This PR moves the consumption of the testing library version (via C++ macro
`SWT_TESTING_LIBRARY_VERSION`) to a C++ function. It seems that defining it
in CMake for C++ targets doesn't trigger for the Swift side when it imports
symbols from C headers.

Will verify with a toolchain build.