Special-case `FileHandle(forWritingAtPath: "CONOUT$")` on Windows.
This PR special-cases calls to `FileHandle.init(forWritingAtPath:)` when passed
`"CONOUT$"` (the reserved "console output" file) on Windows such that it does
not open a new file but just returns a copy of `.stdout` instead.

This change is necessary because, on Windows, opening and locking `"CONOUT$"`
does not lock `stdout` (they're considered entirely different files) and output
written to `"CONOUT$"` is wrapped at the console's column limit even though it's
being written as binary data. The VSCode Swift plugin authors would like to be
able to specify writing the JSON event stream to `"CONOUT$"`, but the observed
behaviour is stopping them from doing so.