Exit tests shouldn't emit crash logs. (#670)
When a process crashes, the platforms we support all write a crash log
(or core, or error report) to disk. This is not necessary for exit test
child processes (since we intentionally crash them.)

This PR suppresses generation of those logs from within an exit test:

- On Darwin, we disable the exception port for crash report files;
- On Linux, we set the maximum core dump size to zero bytes; and
- On Windows, we disable Windows Error Reporting.

The above effects are only applied to the exit test's child process, not
to the parent test process nor the system as a whole.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.