Do not include <sys/resource.h> on WASI
Including `sys/resource.h` from wasi-libc requires a feature macro to be
defined (`-D_WASI_EMULATED_PROCESS_CLOCKS`) and emits compilation errors
otherwise. However, the header is not used in WASI paths, so it is safe
to exclude it from the list of includes.

See `sys/times.h` that is transitively included by `sys/resource.h`
https://github.com/WebAssembly/wasi-libc/blob/wasi-sdk-22/libc-top-half/musl/include/sys/times.h