Require explicit backtrace and source location when creating and recording issues (#688)
This refactors some of the logic for constructing `SourceContext` and
`Issue` instances to require explicit arguments in more places.

### Motivation:

We can provide clearer and more relevant details about recorded issues
by ensuring that their backtrace and/or source location are accurate. In
general, whenever collecting a backtrace, it's better to elide unhelpful
frames from the testing library itself so that frames from user code are
as close to the top of the call stack as possible. Similarly, whenever
an explicit source location is known, it's best to propagate that in any
recorded issues which originate at that location.

Currently, the `SourceContext` type has default parameter values for
both its `backtrace` and `sourceLocation` properties. Having default
values for these parameters means that callers elsewhere in the codebase
may accidentally not specify a value, thereby not propagating useful
information. Or, in the case of backtraces, they may rely on the default
value which results in additional, irrelevant frames in the call stack
(representing thunks inserted by the compiler for the default value).
The initializer for `Issue` also includes a default value for one of its
parameters.

### Modifications:

- Change `SourceContext.init(...)` to remove the default values for both
of its parameters: `backtrace:` and `sourceLocation:`.
- Change `Issue.init(...)` to no longer have default values for any of
its parameters.
- Remove the non-public, static `Issue.record(...)` functions, and
change callers to rely on the instance `record(configuration:)` method.
This consolidates the usage pattern across the `Testing` module for
creating an `Issue` and then recording it.
- Change `SkipInfo.init(...)` to remove the default value for its
`sourceContext:` parameter.
- Add a convenience `init(for:sourceLocation:)` which accepts an `Error`
and correctly creates a `SourceContext` for it.
- Add deprecated overloads of all SPIs which were modified to remove
default parameter values.
- Add convenience overloads to `TestingAdditions.swift` which _preserve_
default parameter values, to avoid unnecessary boilerplate in test-only
code where it's safe to rely on such default values.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.