Hoist JSON Lines logic from EntryPoint.swift.
This PR moves the logic that implements JSON Lines support (i.e. the code that
strips newlines from JSON generated by Foundation) out of EntryPoint.swift so
that it can be used by other callers.

This change was originally part of #697. I'm splitting it out into its own PR to
make that one (both really) easier to read and understand.