Distinguish when an error is part of a recorded issue in an exit test.
If an exit test records an issue of kind `.errorCaught()`, we currently treat it
the same as any other issue and translate it to a flat `.unconditional` issue in
the parent process:

```swift
await #expect(exitsWith: .failure) {
  Issue.record(NoSuchWeaselError()) // unconditional issue recorded
}
```

This PR distinguishes the `.errorCaught` issue kind and records an
`.errorCaught` issue in the parent process. Fidelity of errors across a process
boundary is weak, but we encode a minimal representation of the error that
should be good enough for most uses. If you need `withKnownIssue()`, or
`#expect(throws:)`, you can call them _inside_ the exit test.