Record an issue if an exit test throws an uncaught error.
If an exit test throws an error and doesn't catch it, e.g.:

```swift
await #expect(exitsWith: .failure) {
  throw NoSuchWeaselError()
}
```

Currently, an error thrown in this fashion terminates the exit test's child
process the same way it would if it were thrown uncaught from `main()`.

This PR, in addition to that behavior, records an `.errorCaught` issue in the
parent process. Fidelity of errors across a process boundary is weak, but we
encode a minimal representation of the error that should be good enough for most
uses. If you need `withKnownIssue()`, or `#expect(throws:)`, you can call them
_inside_ the exit test.