Make the `ExitTest` type move-only.
This PR adds `~Copyable` to `ExitTest`, which is a tools-only bit of SPI
representing a (gasp) exit test. This change has the effect of making it
impossible to accidentally invoke an exit test body in the parent process, and
also impossible to accidentally invoke twice in the child process. Phew!