Make the `ExitTest` type move-only. (#712)
This PR adds `~Copyable` to `ExitTest`, which is a tools-only bit of SPI
representing a (gasp) exit test. This change has the effect of making it
impossible to accidentally invoke an exit test body in the parent
process, and also impossible to accidentally invoke twice in the child
process. Phew!

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.