Remove extraneous @escaping and @Sendable attributes from CustomExecutionTrait.execute(...) SPI (#726)
This refines the `CustomExecutionTrait` SPI protocol by removing two
unnecessary attributes from the signature of its
`execute(_:for:testCase:)` method requirement:

1. The `@escaping` attribute on the `function` parameter was needed
originally to avoid a warning, but `function` should _not_ actually be
considered escaping and part of the contract of this interface is that
the implementor must call it before returning. The compiler no longer
complains about this, so we can enforce this requirement by removing the
unnecessary `@escaping` attribute now.
2. The outer `@Sendable` was necessary when this SPI was first written
to avoid a Swift concurrency warning, since a reference to the function
is taken elsewhere in the codebase, but this is no longer needed in
recent Swift compilers.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.