Walk PE files' sections instead of using `swift_enumerateAllMetadataSections()`. (#728)
On Windows, the platform provides sufficient API for us to walk images
at runtime looking for those that contain Swift metadata (similar to
what we do on Darwin.) Therefore, we don't need to use the
`swift_enumerateAllMetadataSections()` function from the Swift runtime
on Windows.

This change also plumbs through an `imageAddress` argument to the
discovery callback function. We're not currently using it, but it should
be useful in the future for diagnostics (e.g. indicating that some data
from a given image is malformed, or for passing to downstream testing
tools that can make use of it.)

This change _also_ also plumbs the `stop` argument _down_ into the
platform implementations so that we don't keep iterating images after
the caller has told us to stop. In practice this has not been a
correctness problem, but it does impact performance of type discovery.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.