[6.0.2][CMake] Explicitly link Testing to Foundation (#713)
Cherry-pick #693 into `release/6.0.2`

* **Explanation**: Previously in CMake builds, when `Foundation` was not
in the regular search directory (e.g. resource directory or SDK search
paths) `#if canImport(Foundation)` used to fail, and the functionalities
are not included. This patch provides a way to provide `Foundation_DIR`
for `find_packgage(Foundation CONFIG)`, so that clients can correctly
link Testing to Foundation
* **Scope**: CMake builds
* **Risk**: Low. No actual code changes.
* **Testing**: Passes current test suite Also manually tested the build
toolchain
* **Issues**: N/A
* **Reviewer**: @stmontgomery

---------

Co-authored-by: Jonathan Grynspan <jgrynspan@apple.com>