Only call `getsectiondata()` once per image.
This PR improves the performance of test discovery by only calling
`getsectiondata()` once per loaded image instead of twice. It also refactors the
C++ platform-specific enumerator function to take a new structure,
`SWTSectionBounds`, instead of multiple arguments. This lets us use a C++
range-for-loop (`for (const auto& record : sectionBounds)`) instead of a raw C
for-loop, reduces register traffic on Windows, and is all around a bit more
object-oriented (which I hope is an improvement, yeah?)

A non-scientific comparison at my desk shows that, before the change, **cold**
test discovery on macOS took an average of 0.0376 seconds. After the change,
the average was 0.0217 seconds for an approximate 1.73x speedup.