Fix macro expansion failure for a non-copyable suite type with a `test` method. (#739)
If a non-copyable suite type has a test function whose name starts with
`test`, we emit a call to `__invokeXCTestCaseMethod()` that's a no-op
(but we can't tell during macro expansion if it will be.) This function
needs to be updated to accept a non-copyable generic type.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.