Simplify the configuration option for synchronous test isolation. (#747)
This PR replaces `Configuration.isMainActorIsolationEnforced` with a new
property `Configuration.defaultIsolationContext` which can be set to any
actor, not just the main actor. This property is then used as the
isolation context for synchronous test functions that are not otherwise
actor-isolated.

The typical use case for this property would be to set it to
`MainActor.shared` so that synchronous test functions run on the main
actor, but a testing tool might want to instead run them all on another
actor or serial executor for performance (or other) reasons.

This change isn't just speculative: it allows us to simplify the
macro-generated thunk for synchronous test functions so that it does not
need to emit the body of the thunk twice (once for the main actor, once
for non-isolated.) This change wasn't possible before `#isolation` was
introduced, but we'd already added
`Configuration.isMainActorIsolationEnforced` when that happened so the
timing didn't line up.

By rewriting this part of the macro expansion, we're able to toss a
bunch of support code that isn't needed anymore. Yay!

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.