Add a return type for exit tests.
This PR adds a new structure, `ExitTest.Result`, that is returned from a call to
`#expect(exitsWith:)` or `#require(exitsWith:)`. It currently contains the
actual exit condition of the test (which should equal the expected exit
condition on success) but in the future we can add additional properties such as
the standard output and standard error streams.